#18 Split base to core and common. Move sequence to core
