+++ base.js
@@ -1,0 +1,5 @@
+'use strict'
+var merge = require('./util/objectMerge')
+module.exports = merge(merge({},
+  require('./core')),
+  require('./common'))
