+++ common/index.js
@@ -1,13 +1,5 @@
 'use strict'
 module.exports = {
-  compose: require('./compose'),
-  isReduced: require('./isReduced'),
-  reduced: require('./reduced'),
-  unreduced: require('./unreduced'),
-  reduce: require('./reduce'),
-  transduce: require('./transduce'),
-  eduction: require('./eduction'),
-  into: require('./into'),
   map: require('./map'),
   filter: require('./filter'),
   remove: require('./remove'),
