+++ common/partitionAll.js
