+++ common/remove.js
