+++ common/take.js
@@ -1,6 +1,6 @@
 'use strict'
 
-var reduced = require('./reduced')
+var reduced = require('../core/reduced')
 
 module.exports =
 function take(n){
