+++ core/compose.js
