+++ core/index.js
@@ -1,0 +1,12 @@
+'use strict'
+module.exports = {
+  compose: require('./compose'),
+  isReduced: require('./isReduced'),
+  reduced: require('./reduced'),
+  unreduced: require('./unreduced'),
+  reduce: require('./reduce'),
+  transduce: require('./transduce'),
+  eduction: require('./eduction'),
+  into: require('./into'),
+  sequence: require('./sequence')
+}
