+++ core/isReduced.js
