+++ core/reduce.js
