+++ core/reduced.js
