+++ core/transduce.js
