+++ iterator/index.js
@@ -6,7 +6,6 @@ module.exports = {
   iterable: require('./iterable'),
   iterator: require('./iterator'),
   toArray: require('./toArray'),
-  sequence: require('./sequence'),
   range: require('./range'),
   count: require('./count'),
   cycle: require('./cycle'),
