+++ string/words.js
@@ -1,5 +1,5 @@
 'use strict'
-var compose = require('../base/compose'),
+var compose = require('../core/compose'),
     isNumber = require('../util/isNumber'),
     split = require('./split'),
     nonEmpty = require('./nonEmpty')
